python/sample-app-webhook/requirements.txt
Django>=5.0.0
djangorestframework>=3.15.0
uvicorn>=0.24.0
python-dotenv>=0.19.0
gohighlevel-api-client==1.0.0b1